Add function descriptions, add dual mobility map
Added function definitions to a few of the functions that have many repeats.

Changed the name of calc_max_mob to calc_dual_max_mob to be consistent.

Created mobility_map_dual, which is a function to plot both the positive and negative voltage sweeps, with max mobility, from the dual max mobility finder.

Created plot_dual_tcurve_fit, which will plot just the same as with plot_tcurve_fit, but it will plot both the positive and negative voltage sweep mobility curves.