Added functions to evaluate hysterisis & transfer curve nonlinearity.
Added calc_hysterisis, which calculates a value for hysterisis to quantify how bad it is.

Added mob_dual_analysis, which calculates the nonlinearity of the transfer curves for both the positive and negative voltage sweeps.

Integrated the maximum mobility calculation into the mob_dual_analysis file.

Created plot_dual_mob_sweep so that you can look at mobility plotted over the course of the sweep. This can also give you a better idea for how bad the curvature of the transfer curve is.