Changed all files so that both p and n type devices can be calculated.
Changed how the porgrams work, so that now the different sweeps are split into forward and backward instead of positive and negative. This allows for the easy application to both p and n type devices.

All functions can now handle p and n type devices, using the input argument 'semiType', and are in terms of forward and backward sweeps, not positive and negative.