Made some visualization updates, also added anisotropy distinguishing things to many functions.
Updated design_space_plot to be a final version of looking at the design space.

Updated Model_Visualization.nb to output a movie of the 3d contour plot.

Updated the analysis functions (calc_avg, mob_dual_analysis, output_stats) to keep track of the different metrics separating for vertical and horizontal channel directions.  Still need to update the optimization functions to handle this.